+++ android/libuv-demo/src/main/cpp/CMakeLists.txt
