+++ android/libuv-demo/src/main/res/mipmap-mdpi/ic_launcher.png
