+++ android/libuv-demo/src/main/res/mipmap-xxhdpi/ic_launcher.png
