+++ android/libuv-demo/src/main/res/mipmap-xxxhdpi/ic_launcher.png
